# docs/swagger.yaml
basePath: /v1
definitions:
...
  domain.Comment:
    properties:
      id:
        readOnly: true
        type: string
      task_id:
        type: string
      text:
        type: string
    required:
    - task_id
    - text
    type: object
...
  domain.Task:
    properties:
      column_id:
        type: string
      description:
        type: string
      id:
        readOnly: true
        type: string
      name:
        type: string
      position:
        type: integer
    required:
    - column_id
    - description
    - name
    type: object
  web.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
info:
  contact: {}
  title: Task Tracker API
  version: "1.0"
paths:
...
  /columns/{id}/tasks:
    get:
      description: get tasks by column id
      parameters:
      - description: column ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get tasks by column id
      tags:
      - columns
  /comments:
    get:
      description: get all comments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Comment'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get all comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: add by json comment
      parameters:
      - description: Add comment
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/domain.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Add a comment
      tags:
      - comments
  /comments/{id}:
    delete:
      description: Delete by comment ID
      parameters:
      - description: comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: it's ok
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Delete a comment
      tags:
      - comments
    get:
      description: get comment by id
      parameters:
      - description: comment ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Comment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Show a comment
      tags:
      - comments
...
  /projects/{id}/tasks:
    get:
      description: get tasks by project id
      parameters:
      - description: project ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get tasks by project id
      tags:
      - projects
  /tasks:
    get:
      description: get all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: add by json task
      parameters:
      - description: Add task
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/domain.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Add a task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete by task ID
      parameters:
      - description: task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: it's ok
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Delete a task
      tags:
      - tasks
    get:
      description: get task by id
      parameters:
      - description: task ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Show a task
      tags:
      - tasks
  /tasks/{id}/comments:
    get:
      description: get tasks by task id
      parameters:
      - description: task ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.HTTPError'
      summary: Get comments by task id
      tags:
      - tasks
swagger: "2.0"